refactor: app.ts에서 데코레이터, 전역 상태, 검증 관련 코드를 파일로 분리
app.ts에서 데코레이터, 전역 상태, 검증 관련 코드를 파일로 분리
